LU-11830 libcfs: quiet print format warning
The watchdog code was reworked in master but generates a warning on
b2_12 due to mismatched print format vs. time64_t.

Test-Parameters: trivial
Reviewed-on: https://review.whamcloud.com/34805
Tested-by: Jenkins
Tested-by: Maloo <maloo@whamcloud.com>

+++ libcfs/libcfs/watchdog.c
@@ -176,9 +176,9 @@ static void lcw_dump_stack(struct lc_watchdog *lcw)
 	delta_time = current_time - lcw_last_watchdog_time;
 	if (delta_time < libcfs_watchdog_ratelimit &&
 	    lcw_recent_watchdog_count > 3) {
-		LCONSOLE_WARN("Service thread pid %u was inactive for %lu.%.02lus. Watchdog stack traces are limited to 3 per %d seconds, skipping this one.\n",
+		LCONSOLE_WARN("Service thread pid %u was inactive for %llu.%.02lus. Watchdog stack traces are limited to 3 per %d seconds, skipping this one.\n",
 			      (int)lcw->lcw_pid,
-			      timediff.tv_sec,
+			      (unsigned long long)timediff.tv_sec,
 			      timediff.tv_nsec / (NSEC_PER_SEC / 100),
 			      libcfs_watchdog_ratelimit);
 	} else {
@@ -190,9 +190,9 @@ static void lcw_dump_stack(struct lc_watchdog *lcw)
 			lcw_recent_watchdog_count = 0;
 		}
 
-		LCONSOLE_WARN("Service thread pid %u was inactive for %lu.%.02lus. The thread might be hung, or it might only be slow and will resume later. Dumping the stack trace for debugging purposes:\n",
+		LCONSOLE_WARN("Service thread pid %u was inactive for %llu.%.02lus. The thread might be hung, or it might only be slow and will resume later. Dumping the stack trace for debugging purposes:\n",
 			      (int)lcw->lcw_pid,
-			      timediff.tv_sec,
+			      (unsigned long long)timediff.tv_sec,
 			      timediff.tv_nsec / (NSEC_PER_SEC / 100));
 		lcw_dump(lcw);
 	}
@@ -391,9 +391,9 @@ static void lcw_update_time(struct lc_watchdog *lcw, const char *message)
 		struct timespec64 timediff;
 
 		timediff = ktime_to_timespec64(lapse);
-		LCONSOLE_WARN("Service thread pid %u %s after %lu.%.02lus. This indicates the system was overloaded (too many service threads, or there were not enough hardware resources).\n",
+		LCONSOLE_WARN("Service thread pid %u %s after %llu.%.02lus. This indicates the system was overloaded (too many service threads, or there were not enough hardware resources).\n",
 			      lcw->lcw_pid, message,
-			      timediff.tv_sec,
+			      (unsigned long long)timediff.tv_sec,
 			      timediff.tv_nsec / (NSEC_PER_SEC / 100));
 	}
 	lcw->lcw_last_touched = newtime;
